One controller for each table

See merge request mabougha/a.d.m.y-teioc!14

+++ src/main/java/dev/astranfalio/teioc/repository/SurveyRepository.java
@@ -4,6 +4,6 @@ import dev.astranfalio.teioc.entity.SurveyEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface SurveyRepository extends JpaRepository<SurveyEntity, Integer> {
+public interface SurveyRepository extends JpaRepository<SurveyEntity, Long> {
 
 }
